+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 <!-- Code generated for API Clients. DO NOT EDIT. -->
+## 0.5.0
+* Renamed `upstream_proto` to `upstream_protocol` for `endpoint` resources
+* Added support for `pooling_enabled` on Endpoints
 
 ## 0.4.0
 * Added support for Cloud Endpoints (currently in private beta).
+* Renamed `principal_id` to `principal` for `endpoint` resources
 
 ## 0.3.0